Update and rename Geometric Mean.cpp to geometric sum of sereies.cpp
